Freeze the semantics streaming would otherwise decide by accident
Six decisions written down as language rather than left as behaviour, before
streaming starts changing what any of them mean.

**A rule is a boundary backtracking does not cross**, and that is §4 now, not a
limitation noted in the status file. A call answers once. Everything else rests
on it: a match records which alternative it came through — one number per call —
which is what lets construction happen after recognition, a fold collect its
steps, and a repetition be committed by `recover`. A rule answering more than
once would be a function to a sequence of matches, and all of those would be
rebuilt around it. The cost is stated too: a grammar cannot be refactored freely.

Writing that section found the example wrong. `Name = "x" | "xy"` has been in
docs/status.md since the beginning as the case that fails, and it matches
perfectly well — the shorter alternative wins and `'y'` takes what is left. The
failing one is `"xy" | "x"`. Two tests now pin both orderings, which is how it
surfaced.

**Nothing is caught.** §8.2 promised an exception out of a `=>` inside a
recovering repetition would be treated as a value failure; the code never did it,
and the specification was the half that was wrong. Catching means catching
`Exception`, because no type separates "this quantity is not a number" from
`NullReferenceException`, and a parser reporting a bug in the author's own C# as
"row 400 was malformed" is worse than one that stops.

**Retention follows from the boundary**, and §6.3 now says so: a call reaches
back not at all, a rule reaches back exactly as far as it has consumed, so what
must be held is the extent of the outermost rule in progress. `recover`'s
synchronization point is a bound by construction. **A capture is materialized
when the rule that owns it accepts**, which is why a capture and a reused buffer
are compatible.

**Absolute offsets are `long`, extents are `int`** — a position is into a file, a
span is into a buffer, and counts are counts. The recovery factory's `position`
and the `OnRecovered` hook widen accordingly.

And §4.5 stops calling itself a working proposal, having been implemented and
tested for some time.

+++ examples/DotGram.Examples/LoggingFeedExample.cs
@@ -92,6 +92,6 @@ public static partial class LoggingFeedReader
 	/// mechanism, and it is why the generated side can afford to always be there.
 	/// </remarks>
 	static partial void OnRecovered(
-		string rule, string text, int position, int line, int column, int ordinal, string message) =>
+		string rule, string text, long position, int line, int column, int ordinal, string message) =>
 		_rejected?.Add(new FeedRejection(rule, line, text, message));
 }
